feat: #39 친구 추가 API 구현
- POST /friend/add 엔드포인트 추가
- friend_id를 받아 친구 관계를 생성
- 성공 시 성공 메시지 반환, 오류 발생 시 오류 메시지 반환

+++ server/userttubeot/src/main/java/com/user/userttubeot/user/domain/entity/FriendId.java
@@ -19,9 +19,9 @@ import lombok.Setter;
 public class FriendId implements Serializable {
 
     @Column(name = "user_id", nullable = false)
-    private Long userId;
+    private Integer userId;
 
     @Column(name = "friend_id", nullable = false)
-    private Long FriendId;
+    private Integer FriendId;
 }
